bs-infra/requirements.txt
Flask>=0.9
dataiku-api-client>=11.0.0